+++ Specifications/Fundamentals/Execution/for_ExecutionContextManager/when_setting_current_statically.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Execution;
+namespace Aksio.Cratis.Execution.for_ExecutionContextManager;
 
 public class when_setting_current_statically : Specification
 {
